feat: Implement Unified File System Abstraction (spec 001)
- Add System.IO.Abstractions (TestableIO) for testable file operations
- Implement server-side SandboxedFileSystem confined to StorageRootPath
- Add PathValidator, FileSizeValidator, ExtensionValidator for security
- Add SHA256 checksum verification for file uploads
- Add file download endpoint with integrity verification
- Fix RpcMessageRegistry static dictionary causing parallel test failures
- Add comprehensive unit and integration tests (465 total)
- Add documentation for IFileSystem usage and configuration

BREAKING: Removed IFileService/LocalDiskFileService - use IFileSystem instead

+++ BitPantry.CommandLine.Remote.SignalR.Client/FileUploadProgress.cs
@@ -1,4 +1,4 @@
 ﻿namespace BitPantry.CommandLine.Remote.SignalR.Client
 {
-    public record FileUploadProgress(int TotalRead, string Error = null) { }
+    public record FileUploadProgress(long TotalRead, string Error = null) { }
 }
